funcionando con codigo separado
