+++ main.cpp
@@ -14,18 +14,17 @@
 #include <sys/types.h>
 #include <semaphore.h>
 #include <chrono>
-#include <sys/resource.h>
 
 using namespace std;
 
-sem_t sem; // semáforo para la cola de genomas aceptados
-string carpeta_genomas; // donde estaran los genomas a procesar
+sem_t sem;
+string carpeta_genomas, tipo; // donde estaran los genomas a procesar
 mutex queueMutex;// mutex para la cola de genomas aceptados
 condition_variable cv; // variable de condicion para la cola de genomas aceptados
 queue<string> genomasAceptados; // esta es la queue que el problema dice q hay q usar pa meter los q cumplen con el umbral
 float umbral; // Umbral para el contenido GC
 bool finalizado = false;
-string tipo;
+vector<thread> threads;
 
 float calcular_promedio_GC(const string& genoma_file) { // calcula el promedio de GC de un genoma
     ifstream genoma_content(carpeta_genomas + genoma_file); // abrir el archivo del genoma (path y nombre del archivo)
@@ -46,8 +45,85 @@ float calcular_promedio_GC(const string& genoma_file) { // calcula el promedio d
     }
 
     return genoma.empty() ? 0 : static_cast<float>(contador_GC) / genoma.size();
-}   // si esta vacio, devuelve cero. Si no lo esta, saca promedio (esto evita que se divida por 0)
+}
 
+void consumir_genomas(string option){
+    if (option == "mutex"){
+        unique_lock<mutex> lock(queueMutex);
+        while (!finalizado || !genomasAceptados.empty()) { 
+            cv.wait(lock, [&]{ return finalizado || !genomasAceptados.empty(); });
+            while (!genomasAceptados.empty()) {
+                cout << "Genoma aceptado: " << genomasAceptados.front() << endl;
+                genomasAceptados.pop();
+            }
+        }
+    }
+    else if (option == "sem"){
+        while (!finalizado || !genomasAceptados.empty()) {
+            sem_wait(&sem); // Espera a que el semáforo esté disponible
+            while (!genomasAceptados.empty()) {
+                cout << "Genoma aceptado: " << genomasAceptados.front() << endl;
+                genomasAceptados.pop();
+            }
+            sem_post(&sem); // Libera el semáforo después de consumir genomas
+        }
+    }
+}
+
+void procesar_genoma(const string& genoma_file, float umbral, string option){
+    if (option == "mutex"){
+        float promedio = calcular_promedio_GC(genoma_file); //calcula el promedio de GC del genoma con la funcion que ya esta definida arriba
+        if (promedio >= umbral) { //si el promedio es mayor o igual al umbral, lo mete a la cola
+            lock_guard<mutex> lock(queueMutex);// lock para la cola de genomas aceptados
+            genomasAceptados.push(genoma_file); //mete el genoma a la cola
+            cv.notify_one();// notifica a la variable de condicion que hay un genoma en la cola
+        }
+    }
+    else if (option == "sem"){
+        float promedio = calcular_promedio_GC(genoma_file);
+        if (promedio >= umbral) {
+            sem_wait(&sem); // Espera a que el semáforo esté disponible
+            genomasAceptados.push(genoma_file);
+            sem_post(&sem); // Libera el semáforo después de agregar el genoma
+        }
+    }
+}
+
+void seleccionMutex(vector<string> genomas){
+    thread consumidor(consumir_genomas,"mutex");//crea el thread que consume los genomas para mutex
+    for (const auto& genoma_file : genomas) { //crea un thread por cada genoma
+        threads.emplace_back(procesar_genoma, genoma_file, umbral,"mutex");//crea el thread que procesa el genoma
+    }
+
+    for (auto& t : threads) {
+        t.join(); //espera a que todos los threads terminen
+    }
+
+    {
+        lock_guard<mutex> lock(queueMutex);
+        finalizado = true;
+    }
+    // Notificar al consumidor después de establecer finalizado en true
+    cv.notify_one();
+    consumidor.join();
+}
+
+void seleccionSemaforo(vector<string> genomas){
+   thread consumidor(consumir_genomas, "sem");//crea el thread que consume los genomas para semaforo
+    for (const auto& genoma_file : genomas) {
+        threads.emplace_back(procesar_genoma, genoma_file, umbral,"sem");
+    }
+
+    for (auto& t : threads) {
+        t.join();
+    }
+
+    sem_wait(&sem); // Espera a que el semáforo esté disponible antes de finalizar
+    finalizado = true;
+    sem_post(&sem); // Libera el semáforo después de establecer finalizado en true
+    consumidor.join();
+    sem_destroy(&sem); // Destruye el semáforo
+}
 vector<string> leer_genomas(const string& carpeta_genomas) { //toma todos los ficheros y retorna un vector con el nombre
                                                             //de todos los archivos en una carpeta
     vector<string> genomas;
@@ -68,67 +144,9 @@ vector<string> leer_genomas(const string& carpeta_genomas) { //toma todos los fi
     return genomas; //retorna el vector con los nombres de los archivos
 }
 
-void procesar_genoma_mutex(const string& genoma_file, float umbral) { //procesa un genoma y lo mete a la cola si cumple con el umbral
-    float promedio = calcular_promedio_GC(genoma_file); //calcula el promedio de GC del genoma con la funcion que ya esta definida arriba
-    if (promedio >= umbral) { //si el promedio es mayor o igual al umbral, lo mete a la cola
-        lock_guard<mutex> lock(queueMutex);// lock para la cola de genomas aceptados
-        genomasAceptados.push(genoma_file); //mete el genoma a la cola
-        cv.notify_one();// notifica a la variable de condicion que hay un genoma en la cola
-    } //si no cumple con el umbral, no hace nada
-    //importante: "umbral" es un parametro que pasar por consola ****PENDIENTE****
-}
-
-void consumir_genomas_mutex() {// funcion que consume los genomas de la cola
-    cout<<"consumidor iniciado"<<endl;
-    unique_lock<mutex> lock(queueMutex);  // lock para la cola de genomas aceptados
-    while (!finalizado || !genomasAceptados.empty()) { //mientras no se haya finalizado o la cola no este vacia
-        cv.wait(lock, [&]{ return finalizado || !genomasAceptados.empty(); }); //espera a que haya un genoma en la cola o se finalice
-        while (!genomasAceptados.empty()) { //mientras la cola no este vacia
-            cout << "Genoma aceptado: " << genomasAceptados.front() << endl; //saca el genoma de la cola
-            genomasAceptados.pop();// lo saca de la cola (pop)
-            cout<<"while interno"<<endl;
-        }
-        cout<<"primer while"<<endl;
-    }
-    cout<<"consumidor finalizado"<<endl;
-}
-
-void procesar_genoma_semaforo(const string& genoma_file, float umbral) {
-    float promedio = calcular_promedio_GC(genoma_file);
-    if (promedio >= umbral) {
-        sem_wait(&sem); // Espera a que el semáforo esté disponible
-        genomasAceptados.push(genoma_file);
-        sem_post(&sem); // Libera el semáforo después de agregar el genoma
-    }
-}
-
-
-void consumir_genomas_semaforo() {
-    //cout << "consumidor iniciado" << endl;
-    while (!finalizado || !genomasAceptados.empty()) {
-        sem_wait(&sem); // Espera a que el semáforo esté disponible
-        while (!genomasAceptados.empty()) {
-            cout << "Genoma aceptado: " << genomasAceptados.front() << endl;
-            genomasAceptados.pop();
-            //cout << "while interno" << endl;
-        }
-        sem_post(&sem); // Libera el semáforo después de consumir genomas
-        //cout << "primer while" << endl;
-    }
-    //cout << "consumidor finalizado" << endl;
-}
-
-void measureMemoryUsage() {
-    struct rusage usage;
-    getrusage(RUSAGE_SELF, &usage);
-
-    std::cout << "Memory used: " << usage.ru_maxrss << " kilobytes\n";
-}
-
 int main(int argc, char* argv[]) {
-    auto start = std::chrono::high_resolution_clock::now(); // Start time
-    if (argc < 4) { 
-            cerr << "Error: No se proporcionaron los parametros necesarios." << endl;
+    if (argc < 3) { // verifica que se haya pasado al menos un argumento
+            cerr << "Error: No se proporcionó un umbral." << endl;
             return 1;
     }
 
@@ -141,54 +159,14 @@ int main(int argc, char* argv[]) {
     sem_init(&sem, 0, 1); // Inicializa el semáforo
     thread consumidor;
     if(tipo == "mutex"){
-        thread consumidor(consumir_genomas_mutex);//crea el thread que consume los genomas para mutex
-    for (const auto& genoma_file : genomas) { //crea un thread por cada genoma
-        cout<<"t"<<endl;
-        threads.emplace_back(procesar_genoma_mutex, genoma_file, umbral);//crea el thread que procesa el genoma
-    }
-
-    for (auto& t : threads) {
-        cout<<"esperando"<<endl;
-        t.join(); //espera a que todos los threads terminen
-    }
-
-    {
-        lock_guard<mutex> lock(queueMutex);
-        cout << "finalizado" << endl;
-        finalizado = true;
-    }
-    // Notificar al consumidor después de establecer finalizado en true
-    cv.notify_one();
+        seleccionMutex(genomas);
     }
     else if(tipo == "sem"){
-        thread consumidor(consumir_genomas_semaforo);//crea el thread que consume los genomas para semaforo
-        for (const auto& genoma_file : genomas) {
-        threads.emplace_back(procesar_genoma_semaforo, genoma_file, umbral);
-    }
-
-    for (auto& t : threads) {
-        cout << "esperando" << endl;
-        t.join();
-    }
-
-    sem_wait(&sem); // Espera a que el semáforo esté disponible antes de finalizar
-    finalizado = true;
-    sem_post(&sem); // Libera el semáforo después de establecer finalizado en true
-    consumidor.join();
-    sem_destroy(&sem); // Destruye el semáforo
+        seleccionSemaforo(genomas);
     }
     else{
         cout << "Error en el tercer argumento" << endl;
         return 0;
     }
-
-    consumidor.join(); //espera a que el consumidor termine ACA ESTA MURIENDO ***ARREGLAR***
-    cout<<"consumidor"<<endl;
-     auto end = std::chrono::high_resolution_clock::now(); // End time
-
-        // Calculate and print the time taken
-        std::chrono::duration<double> elapsed = end - start;
-        std::cout << "Time taken: " << elapsed.count() << " seconds\n";
-        measureMemoryUsage();
     return 0;
 }
